Some `quota` package cleanups
