+++ pkg/sys/fs/xfs/quota/quota_linux.go
@@ -145,13 +145,3 @@ func SetQuota(ctx context.Context, path, volumeID, blockFile string, quota FSQuo
 		"ProjectID", projectID)
 	return nil
 }
-
-type DefaultDriveQuotaer struct{}
-
-func (q *DefaultDriveQuotaer) SetQuota(ctx context.Context, path, volumeID, blockFile string, quota FSQuota) error {
-	return SetQuota(ctx, path, volumeID, blockFile, quota)
-}
-
-func (q *DefaultDriveQuotaer) GetQuota(blockFile, volumeID string) (FSQuota, error) {
-	return GetQuota(blockFile, volumeID)
-}
